# api/src/domain/models.ts
export interface User {
  id: string;
  username: string;
  normalizedUsername: string;
  passwordHash: string;
  role: UserRole;
  createdAt: Date;
  updatedAt: Date;
}

export type UserRole = 'admin' | 'teacher';

export interface TeacherAccount {
  id: string;
  username: string;
  normalizedUsername: string;
  role: 'teacher';
  roomIds: string[];
  createdAt: Date;
  updatedAt: Date;
}

export interface Session {
  id: string;
  userId: string;
  tokenHash: string;
  csrfHash: string;
  expiresAt: Date;
  createdAt: Date;
}

export interface NewsItem {
  id: string;
  text: string;
  position: number;
  createdAt: Date;
}

export interface Program {
  id: string;
  title: string;
  description: string;
  video: string;
  category: string | null;
  position: number;
  createdAt: Date;
}

export type PrivateRoomSourceType = 'live' | 'youtube' | 'video' | 'external';
export type PrivateRoomSupportMaterialType = 'url' | 'image' | 'pdf';

export interface PrivateRoom {
  id: string;
  roomCode: string;
  title: string;
  description: string;
  sourceType: PrivateRoomSourceType;
  sourceUrl: string;
  supportMaterialEnabled: boolean;
  supportMaterialTitle: string;
  supportMaterialType: PrivateRoomSupportMaterialType;
  supportMaterialUrl: string;
  supportMaterialCurrentPage: number;
  isActive: boolean;
  expiresAt?: Date | null;
  createdAt: Date;
  updatedAt: Date;
}

export interface PrivateRoomAccessSession {
  id: string;
  roomId: string;
  tokenHash: string;
  expiresAt: Date;
  createdAt: Date;
}

export type PrivateRoomInteractionMode = 'questions_comments' | 'questions_only' | 'comments_only';
export type PrivateRoomMessageStatus = 'pending' | 'approved' | 'hidden' | 'answered' | 'archived';

export interface PrivateRoomInteractionSettings {
  roomId: string;
  enabled: boolean;
  mode: PrivateRoomInteractionMode;
  requireName: boolean;
  allowAnonymous: boolean;
  collectContact: boolean;
  moderationRequired: boolean;
  allowPublicReplies: boolean;
  noticeText: string;
  updatedAt: Date;
}

export interface PrivateRoomMessage {
  id: string;
  roomId: string;
  participantName: string;
  participantContact: string;
  body: string;
  adminReply: string;
  status: PrivateRoomMessageStatus;
  isHighlighted: boolean;
  ipHash?: string | null;
  userAgent: string;
  moderatedBy?: string | null;
  moderatedAt?: Date | null;
  createdAt: Date;
  updatedAt: Date;
}

export interface Branding {
  companyName: string;
  tagline: string;
  watermarkText: string;
  logoText: string;
  logoUrl: string;
  backgroundUrl: string;
  scheduleTitle: string;
  tickerLabel: string;
  rssNewsUrl: string;
  partnerLabel: string;
  liveSource: 'obs' | 'youtube';
  liveYoutubeUrl: string;
  liveTitle: string;
  liveDescription: string;
  loopTitle: string;
  loopDescription: string;
  legalName: string;
  legalEmail: string;
  legalCnpj: string;
  legalCity: string;
  legalPhone: string;
  updatedAt: Date;
}

export interface HeaderLink {
  id: string;
  name: string;
  url: string;
  position: number;
  createdAt: Date;
}

export interface Partner {
  id: string;
  name: string;
  logoUrl: string;
  destinationUrl: string;
  position: number;
  createdAt: Date;
}

export interface AuditLog {
  id: string;
  actorUserId?: string;
  action: string;
  targetType: string;
  targetId?: string;
  requestId?: string;
  ip?: string;
  metadata: Record<string, unknown>;
  createdAt: Date;
}

export type MediaKind = 'image' | 'video' | 'document';

export interface MediaAsset {
  id: string;
  kind: MediaKind;
  storageKey: string;
  mimeType: string;
  byteSize: number;
  sha256: string;
  originalName: string;
  createdBy: string;
  createdAt: Date;
}

export const DEFAULT_BRANDING: Omit<Branding, 'updatedAt'> = {
  companyName: 'TV Carlos',
  tagline: 'SINAL INDEPENDENTE · BRASIL',
  watermarkText: 'TV CARLOS • CONTEÚDO EXCLUSIVO',
  logoText: 'TV CARLOS',
  logoUrl: '',
  backgroundUrl: '',
  scheduleTitle: 'Próximos vídeos',
  tickerLabel: 'GIRO TVC',
  rssNewsUrl: 'https://api.rss2json.com/v1/api.json?rss_url=https://g1.globo.com/rss/g1/',
  partnerLabel: 'PARCEIRO',
  liveSource: 'obs',
  liveYoutubeUrl: '',
  liveTitle: 'Transmissão Especial Ao Vivo',
  liveDescription: 'Transmissão em tempo real',
  loopTitle: 'Programação Linear 24h',
  loopDescription: 'TV Carlos - Transmissão Automática',
  legalName: 'Carlos Antonio de Oliveira Piquet',
  legalEmail: 'carlos.piquet2016@gmail.com',
  legalCnpj: '27.658.099/0001-70',
  legalCity: 'Rio de Janeiro - RJ',
  legalPhone: '+55 21 97905-4104',
};