# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TV Carlos API
  version: 3.0.0
  description: API pública e administrativa da Web TV. Escritas exigem cookie de sessão e token CSRF.
servers:
  - url: http://localhost:8082
    description: Docker local
...
security: []
paths:
...
  /api/operations/status:
    get:
      tags: [Operations]
      operationId: getOperationsStatus
      security: [{ sessionCookie: [] }]
      responses:
        '200': { description: Monitor operacional usado pelo dashboard principal, content: { application/json: { schema: { $ref: '#/components/schemas/OperationsStatus' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/auth/login:
    post:
      tags: [Auth]
      operationId: login
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/LoginInput' } } }
      responses:
        '200': { description: Sessão criada; envia cookies tv_session e tv_csrf., content: { application/json: { schema: { $ref: '#/components/schemas/Session' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/auth/session:
    get:
      tags: [Auth]
      operationId: getSession
      security: [{ sessionCookie: [] }]
      responses:
        '200': { description: Sessão atual, content: { application/json: { schema: { $ref: '#/components/schemas/Session' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      security: [{ sessionCookie: [], csrfHeader: [] }]
      responses: { '204': { description: Sessão revogada }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' } }
  /api/auth/credentials:
    put:
      tags: [Auth]
      operationId: changeCredentials
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CredentialsInput' } } }
      responses: { '204': { description: Credenciais alteradas e sessões revogadas }, '400': { $ref: '#/components/responses/ValidationError' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' } }
  /api/teachers:
    get:
      tags: [Auth]
      operationId: listTeachers
      security: [{ sessionCookie: [] }]
      responses: { '200': { description: Professores convidados e salas atribuídas, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/TeacherAccount' } } } } }, '403': { $ref: '#/components/responses/Forbidden' } }
    post:
      tags: [Auth]
      operationId: createTeacher
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TeacherCreateInput' } } } }
      responses: { '201': { description: Professor criado; a senha só aparece nesta resposta, content: { application/json: { schema: { $ref: '#/components/schemas/CreateTeacherResponse' } } } }, '400': { $ref: '#/components/responses/ValidationError' }, '403': { $ref: '#/components/responses/Forbidden' } }
  /api/teachers/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    put:
      tags: [Auth]
      operationId: updateTeacherRooms
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TeacherUpdateInput' } } } }
      responses: { '200': { description: Salas do professor atualizadas, content: { application/json: { schema: { $ref: '#/components/schemas/TeacherAccount' } } } }, '404': { $ref: '#/components/responses/NotFound' }, '403': { $ref: '#/components/responses/Forbidden' } }
    delete:
      tags: [Auth]
      operationId: deleteTeacher
      security: [{ sessionCookie: [], csrfHeader: [] }]
      responses: { '204': { description: Professor removido e sessões associadas revogadas }, '404': { $ref: '#/components/responses/NotFound' }, '403': { $ref: '#/components/responses/Forbidden' } }
  /api/teachers/{id}/rotate-password:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [Auth]
      operationId: rotateTeacherPassword
      security: [{ sessionCookie: [], csrfHeader: [] }]
      responses: { '200': { description: Nova senha do professor gerada, content: { application/json: { schema: { $ref: '#/components/schemas/CreateTeacherResponse' } } } }, '404': { $ref: '#/components/responses/NotFound' }, '403': { $ref: '#/components/responses/Forbidden' } }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: tv_session }
    privateRoomCookie: { type: apiKey, in: cookie, name: tv_private_room_session }
    csrfHeader: { type: apiKey, in: header, name: X-CSRF-Token }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
    ValidationError: { description: Entrada rejeitada, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unauthorized: { description: Sessão ausente ou expirada, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Forbidden: { description: Origem ou CSRF rejeitado, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    NotFound: { description: Recurso inexistente, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    RateLimited: { description: Limite de requisições excedido, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
...
    ServiceHealthStatus: { type: string, enum: [ok, warning, error, neutral] }
    OperationalService:
      type: object
      required: [id, label, status, detail, checkedAt]
      properties:
        id: { type: string }
        label: { type: string }
        status: { $ref: '#/components/schemas/ServiceHealthStatus' }
        detail: { type: string }
        checkedAt: { type: string, format: date-time }
        metadata: { type: object }
    OperationsStatus:
      type: object
      required: [checkedAt, status, summary, services, logs]
      properties:
        checkedAt: { type: string, format: date-time }
        status: { $ref: '#/components/schemas/ServiceHealthStatus' }
        summary:
          type: object
          required: [ok, warning, error, neutral]
          properties: { ok: { type: integer }, warning: { type: integer }, error: { type: integer }, neutral: { type: integer } }
        services: { type: array, items: { $ref: '#/components/schemas/OperationalService' } }
        logs: { type: array, items: { $ref: '#/components/schemas/AuditLog' } }
    Error:
      type: object
      required: [error]
      properties:
        error:
            type: object
            required: [code, message]
            properties: { code: { type: string }, message: { type: string } }
    LoginInput: { type: object, additionalProperties: false, required: [username, password], properties: { username: { type: string, minLength: 3, maxLength: 80 }, password: { type: string, minLength: 1, maxLength: 128, format: password } } }
    CredentialsInput: { type: object, additionalProperties: false, required: [currentPassword, newUsername, newPassword], properties: { currentPassword: { type: string, format: password }, newUsername: { type: string, minLength: 3, maxLength: 80 }, newPassword: { type: string, minLength: 14, maxLength: 128, format: password } } }
    Session: { type: object, required: [user, expiresAt], properties: { user: { type: object, required: [id, username, role], properties: { id: { type: string, format: uuid }, username: { type: string }, role: { type: string, enum: [admin, teacher] } } }, expiresAt: { type: string, format: date-time } } }
    TeacherCreateInput: { type: object, additionalProperties: false, required: [username], properties: { username: { type: string, minLength: 3, maxLength: 80 }, roomIds: { type: array, maxItems: 100, items: { type: string, format: uuid } } } }
    TeacherUpdateInput: { type: object, additionalProperties: false, required: [roomIds], properties: { roomIds: { type: array, maxItems: 100, items: { type: string, format: uuid } } } }
    TeacherAccount:
      type: object
      required: [id, username, normalizedUsername, role, roomIds, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        normalizedUsername: { type: string }
        role: { type: string, enum: [teacher] }
        roomIds: { type: array, items: { type: string, format: uuid } }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    TeacherSummary:
      type: object
      required: [id, username, role]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        role: { type: string, enum: [teacher] }
    CreateTeacherResponse:
      type: object
      additionalProperties: false
      required: [teacher, password]
      properties:
        teacher: { anyOf: [{ $ref: '#/components/schemas/TeacherAccount' }, { $ref: '#/components/schemas/TeacherSummary' }] }
        password: { type: string, format: password, description: "Senha exibida apenas na criação/rotação." }
...
    AuditLog: { type: object, required: [id, action, targetType, metadata, createdAt], properties: { id: { type: string, format: uuid }, action: { type: string }, targetType: { type: string }, targetId: { type: [string, 'null'] }, requestId: { type: [string, 'null'] }, ip: { type: [string, 'null'] }, metadata: { type: object }, createdAt: { type: string, format: date-time } } }
...